Fix link parsing
In the previous iteration of the link parsing algorithm, it searched for
a big section at the end and then cut it off (which may not be ideal).
This means that if there are no links then it will accidentally cut off
the last markdown section header.

The new version works by searching for a markdown link string (a version
number block followed by a link) on each newline and then removing that
newline from the overall data section. Meaning that only the markdown
link blocks are removed (and removing nothing if there are no links).